[DOK-143] refactor : 파일구조 및 코드 수정
